Move scenery tool functions from TopToolbar.cpp into Scenery.cpp
