+++ src/openrct2-ui/windows/Scenery.cpp
@@ -16,7 +16,16 @@
 #include <openrct2/GameState.h>
 #include <openrct2/Input.h>
 #include <openrct2/OpenRCT2.h>
+#include <openrct2/actions/BannerPlaceAction.h>
+#include <openrct2/actions/BannerSetColourAction.h>
+#include <openrct2/actions/FootpathAdditionPlaceAction.h>
+#include <openrct2/actions/LargeSceneryPlaceAction.h>
+#include <openrct2/actions/LargeScenerySetColourAction.h>
 #include <openrct2/actions/ScenerySetRestrictedAction.h>
+#include <openrct2/actions/SmallSceneryPlaceAction.h>
+#include <openrct2/actions/SmallScenerySetColourAction.h>
+#include <openrct2/actions/WallPlaceAction.h>
+#include <openrct2/actions/WallSetColourAction.h>
 #include <openrct2/audio/audio.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/core/Guard.hpp>
@@ -33,6 +42,7 @@
 #include <openrct2/object/SceneryGroupEntry.h>
 #include <openrct2/object/SmallSceneryEntry.h>
 #include <openrct2/object/WallSceneryEntry.h>
+#include <openrct2/paint/VirtualFloor.h>
 #include <openrct2/sprites.h>
 #include <openrct2/world/Park.h>
 #include <openrct2/world/Scenery.h>
@@ -74,6 +84,7 @@ namespace OpenRCT2::Ui::Windows
         WIDX_SCENERY_TAB_1,
     };
 
+    validate_global_widx(WC_SCENERY, WIDX_SCENERY_BACKGROUND);
     validate_global_widx(WC_SCENERY, WIDX_SCENERY_TAB_1);
     validate_global_widx(WC_SCENERY, WIDX_SCENERY_ROTATE_OBJECTS_BUTTON);
     validate_global_widx(WC_SCENERY, WIDX_SCENERY_EYEDROPPER_BUTTON);
@@ -181,6 +192,9 @@ static Widget WindowSceneryBaseWidgets[] = {
         ScenerySelection _selectedScenery;
         int16_t _hoverCounter;
         SceneryTabInfo _filteredSceneryTab;
+
+        uint8_t _unkF64F0E{ 0 };
+        int16_t _unkF64F0A{ 0 };
 
     public:
         void OnOpen() override
@@ -848,6 +862,39 @@ static Widget WindowSceneryBaseWidgets[] = {
             if (scrollIndex == SceneryContentScrollIndex)
             {
                 ContentScrollDraw(dpi);
+            }
+        }
+
+        void OnToolUpdate(WidgetIndex widgetIndex, const ScreenCoordsXY& screenCoords) override
+        {
+            switch (widgetIndex)
+            {
+                case WIDX_SCENERY_BACKGROUND:
+                    ToolUpdateScenery(screenCoords);
+                    break;
+            }
+        }
+
+        void OnToolDown(WidgetIndex widgetIndex, const ScreenCoordsXY& screenCoords) override
+        {
+            switch (widgetIndex)
+            {
+                case WIDX_SCENERY_BACKGROUND:
+                    SceneryToolDown(screenCoords, widgetIndex);
+                    break;
+            }
+        }
+
+        void OnToolDrag(WidgetIndex widgetIndex, const ScreenCoordsXY& screenCoords) override
+        {
+            switch (widgetIndex)
+            {
+                case WIDX_SCENERY_BACKGROUND:
+                    if (gWindowSceneryPaintEnabled & 1)
+                        SceneryToolDown(screenCoords, widgetIndex);
+                    if (gWindowSceneryEyedropperEnabled)
+                        SceneryToolDown(screenCoords, widgetIndex);
+                    break;
             }
         }
 
@@ -1684,6 +1731,1466 @@ static Widget WindowSceneryBaseWidgets[] = {
                 }
             }
         }
+
+        /**
+         *
+         *  rct2: 0x006E287B
+         */
+        void ToolUpdateScenery(const ScreenCoordsXY& screenPos)
+        {
+            MapInvalidateSelectionRect();
+            MapInvalidateMapSelectionTiles();
+
+            if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+            {
+                VirtualFloorInvalidate();
+            }
+
+            gMapSelectFlags &= ~MAP_SELECT_FLAG_ENABLE;
+            gMapSelectFlags &= ~MAP_SELECT_FLAG_ENABLE_CONSTRUCT;
+
+            if (gWindowSceneryPaintEnabled)
+                return;
+            if (gWindowSceneryEyedropperEnabled)
+                return;
+
+            const auto selection = WindowSceneryGetTabSelection();
+            if (selection.IsUndefined())
+            {
+                SceneryRemoveGhostToolPlacement();
+                return;
+            }
+
+            money64 cost = 0;
+
+            switch (selection.SceneryType)
+            {
+                case SCENERY_TYPE_SMALL:
+                {
+                    CoordsXY mapTile = {};
+                    uint8_t quadrant;
+                    Direction rotation;
+
+                    Sub6E1F34SmallScenery(screenPos, selection.EntryIndex, mapTile, &quadrant, &rotation);
+
+                    if (mapTile.IsNull())
+                    {
+                        SceneryRemoveGhostToolPlacement();
+                        return;
+                    }
+
+                    gMapSelectFlags |= MAP_SELECT_FLAG_ENABLE;
+                    if (gWindowSceneryScatterEnabled)
+                    {
+                        uint16_t cluster_size = (gWindowSceneryScatterSize - 1) * kCoordsXYStep;
+                        gMapSelectPositionA.x = mapTile.x - cluster_size / 2;
+                        gMapSelectPositionA.y = mapTile.y - cluster_size / 2;
+                        gMapSelectPositionB.x = mapTile.x + cluster_size / 2;
+                        gMapSelectPositionB.y = mapTile.y + cluster_size / 2;
+                        if (gWindowSceneryScatterSize % 2 == 0)
+                        {
+                            gMapSelectPositionB.x += kCoordsXYStep;
+                            gMapSelectPositionB.y += kCoordsXYStep;
+                        }
+                    }
+                    else
+                    {
+                        gMapSelectPositionA.x = mapTile.x;
+                        gMapSelectPositionA.y = mapTile.y;
+                        gMapSelectPositionB.x = mapTile.x;
+                        gMapSelectPositionB.y = mapTile.y;
+                    }
+
+                    auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<SmallSceneryEntry>(selection.EntryIndex);
+
+                    gMapSelectType = MAP_SELECT_TYPE_FULL;
+                    if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_FULL_TILE) && !gWindowSceneryScatterEnabled)
+                    {
+                        gMapSelectType = MAP_SELECT_TYPE_QUARTER_0 + (quadrant ^ 2);
+                    }
+
+                    MapInvalidateSelectionRect();
+
+                    // If no change in ghost placement
+                    if ((gSceneryGhostType & SCENERY_GHOST_FLAG_0) && mapTile == gSceneryGhostPosition && quadrant == _unkF64F0E
+                        && gSceneryPlaceZ == _unkF64F0A && gSceneryPlaceObject.SceneryType == SCENERY_TYPE_SMALL
+                        && gSceneryPlaceObject.EntryIndex == selection.EntryIndex)
+                    {
+                        return;
+                    }
+
+                    SceneryRemoveGhostToolPlacement();
+
+                    _unkF64F0E = quadrant;
+                    _unkF64F0A = gSceneryPlaceZ;
+
+                    uint8_t attemptsLeft = 1;
+                    if (gSceneryPlaceZ != 0 && gSceneryShiftPressed)
+                    {
+                        attemptsLeft = 20;
+                    }
+
+                    for (; attemptsLeft != 0; attemptsLeft--)
+                    {
+                        cost = TryPlaceGhostSmallScenery(
+                            { mapTile, gSceneryPlaceZ, rotation }, quadrant, selection.EntryIndex, gWindowSceneryPrimaryColour,
+                            gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                        if (cost != kMoney64Undefined)
+                            break;
+                        gSceneryPlaceZ += 8;
+                    }
+
+                    gSceneryPlaceCost = cost;
+                    break;
+                }
+                case SCENERY_TYPE_PATH_ITEM:
+                {
+                    CoordsXY mapTile = {};
+                    int32_t z;
+
+                    Sub6E1F34PathItem(screenPos, selection.EntryIndex, mapTile, &z);
+
+                    if (mapTile.IsNull())
+                    {
+                        SceneryRemoveGhostToolPlacement();
+                        return;
+                    }
+
+                    gMapSelectFlags |= MAP_SELECT_FLAG_ENABLE;
+                    gMapSelectPositionA.x = mapTile.x;
+                    gMapSelectPositionA.y = mapTile.y;
+                    gMapSelectPositionB.x = mapTile.x;
+                    gMapSelectPositionB.y = mapTile.y;
+                    gMapSelectType = MAP_SELECT_TYPE_FULL;
+
+                    MapInvalidateSelectionRect();
+
+                    // If no change in ghost placement
+                    if ((gSceneryGhostType & SCENERY_GHOST_FLAG_1) && mapTile == gSceneryGhostPosition
+                        && z == gSceneryGhostPosition.z)
+                    {
+                        return;
+                    }
+
+                    SceneryRemoveGhostToolPlacement();
+
+                    cost = TryPlaceGhostPathAddition({ mapTile, z }, selection.EntryIndex);
+
+                    gSceneryPlaceCost = cost;
+                    break;
+                }
+                case SCENERY_TYPE_WALL:
+                {
+                    CoordsXY mapTile = {};
+                    uint8_t edge;
+
+                    Sub6E1F34Wall(screenPos, selection.EntryIndex, mapTile, &edge);
+
+                    if (mapTile.IsNull())
+                    {
+                        SceneryRemoveGhostToolPlacement();
+                        return;
+                    }
+
+                    gMapSelectFlags |= MAP_SELECT_FLAG_ENABLE;
+                    gMapSelectPositionA.x = mapTile.x;
+                    gMapSelectPositionA.y = mapTile.y;
+                    gMapSelectPositionB.x = mapTile.x;
+                    gMapSelectPositionB.y = mapTile.y;
+                    gMapSelectType = MAP_SELECT_TYPE_EDGE_0 + edge;
+
+                    MapInvalidateSelectionRect();
+
+                    // If no change in ghost placement
+                    if ((gSceneryGhostType & SCENERY_GHOST_FLAG_2) && mapTile == gSceneryGhostPosition
+                        && edge == gSceneryGhostWallRotation && gSceneryPlaceZ == _unkF64F0A)
+                    {
+                        return;
+                    }
+
+                    SceneryRemoveGhostToolPlacement();
+
+                    gSceneryGhostWallRotation = edge;
+                    _unkF64F0A = gSceneryPlaceZ;
+
+                    uint8_t attemptsLeft = 1;
+                    if (gSceneryPlaceZ != 0 && gSceneryShiftPressed)
+                    {
+                        attemptsLeft = 20;
+                    }
+
+                    cost = 0;
+                    for (; attemptsLeft != 0; attemptsLeft--)
+                    {
+                        cost = TryPlaceGhostWall(
+                            { mapTile, gSceneryPlaceZ }, edge, selection.EntryIndex, gWindowSceneryPrimaryColour,
+                            gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                        if (cost != kMoney64Undefined)
+                            break;
+                        gSceneryPlaceZ += 8;
+                    }
+
+                    gSceneryPlaceCost = cost;
+                    break;
+                }
+                case SCENERY_TYPE_LARGE:
+                {
+                    CoordsXY mapTile = {};
+                    Direction direction;
+
+                    Sub6E1F34LargeScenery(screenPos, selection.EntryIndex, mapTile, &direction);
+
+                    if (mapTile.IsNull())
+                    {
+                        SceneryRemoveGhostToolPlacement();
+                        return;
+                    }
+
+                    auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<LargeSceneryEntry>(selection.EntryIndex);
+                    gMapSelectionTiles.clear();
+
+                    for (auto* tile = sceneryEntry->tiles;
+                         tile->x_offset != static_cast<int16_t>(static_cast<uint16_t>(0xFFFF)); tile++)
+                    {
+                        CoordsXY tileLocation = { tile->x_offset, tile->y_offset };
+                        auto rotatedTileCoords = tileLocation.Rotate(direction);
+
+                        rotatedTileCoords.x += mapTile.x;
+                        rotatedTileCoords.y += mapTile.y;
+
+                        gMapSelectionTiles.push_back(rotatedTileCoords);
+                    }
+
+                    gMapSelectFlags |= MAP_SELECT_FLAG_ENABLE_CONSTRUCT;
+                    MapInvalidateMapSelectionTiles();
+
+                    // If no change in ghost placement
+                    if ((gSceneryGhostType & SCENERY_GHOST_FLAG_3) && mapTile == gSceneryGhostPosition
+                        && gSceneryPlaceZ == _unkF64F0A && gSceneryPlaceObject.SceneryType == SCENERY_TYPE_LARGE
+                        && gSceneryPlaceObject.EntryIndex == selection.EntryIndex)
+                    {
+                        return;
+                    }
+
+                    SceneryRemoveGhostToolPlacement();
+
+                    gSceneryPlaceObject.SceneryType = SCENERY_TYPE_LARGE;
+                    gSceneryPlaceObject.EntryIndex = selection.EntryIndex;
+                    _unkF64F0A = gSceneryPlaceZ;
+
+                    uint8_t attemptsLeft = 1;
+                    if (gSceneryPlaceZ != 0 && gSceneryShiftPressed)
+                    {
+                        attemptsLeft = 20;
+                    }
+
+                    cost = 0;
+                    for (; attemptsLeft != 0; attemptsLeft--)
+                    {
+                        cost = TryPlaceGhostLargeScenery(
+                            { mapTile, gSceneryPlaceZ, direction }, selection.EntryIndex, gWindowSceneryPrimaryColour,
+                            gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                        if (cost != kMoney64Undefined)
+                            break;
+                        gSceneryPlaceZ += kCoordsZStep;
+                    }
+
+                    gSceneryPlaceCost = cost;
+                    break;
+                }
+                case SCENERY_TYPE_BANNER:
+                {
+                    CoordsXY mapTile = {};
+                    Direction direction;
+                    int32_t z;
+
+                    Sub6E1F34Banner(screenPos, selection.EntryIndex, mapTile, &z, &direction);
+
+                    if (mapTile.IsNull())
+                    {
+                        SceneryRemoveGhostToolPlacement();
+                        return;
+                    }
+
+                    gMapSelectFlags |= MAP_SELECT_FLAG_ENABLE;
+                    gMapSelectPositionA.x = mapTile.x;
+                    gMapSelectPositionA.y = mapTile.y;
+                    gMapSelectPositionB.x = mapTile.x;
+                    gMapSelectPositionB.y = mapTile.y;
+                    gMapSelectType = MAP_SELECT_TYPE_FULL;
+
+                    MapInvalidateSelectionRect();
+
+                    // If no change in ghost placement
+                    if ((gSceneryGhostType & SCENERY_GHOST_FLAG_4) && mapTile == gSceneryGhostPosition
+                        && z == gSceneryGhostPosition.z && direction == gSceneryPlaceRotation)
+                    {
+                        return;
+                    }
+
+                    SceneryRemoveGhostToolPlacement();
+
+                    cost = TryPlaceGhostBanner({ mapTile, z, direction }, selection.EntryIndex);
+
+                    gSceneryPlaceCost = cost;
+                    break;
+                }
+            }
+        }
+
+        /**
+         *
+         *  rct2: 0x006E24F6
+         * On failure returns kMoney64Undefined
+         * On success places ghost scenery and returns cost to place proper
+         */
+        money64 TryPlaceGhostSmallScenery(
+            CoordsXYZD loc, uint8_t quadrant, ObjectEntryIndex entryIndex, colour_t primaryColour, colour_t secondaryColour,
+            colour_t tertiaryColour)
+        {
+            SceneryRemoveGhostToolPlacement();
+
+            // 6e252b
+            auto smallSceneryPlaceAction = SmallSceneryPlaceAction(
+                loc, quadrant, entryIndex, primaryColour, secondaryColour, tertiaryColour);
+            smallSceneryPlaceAction.SetFlags(GAME_COMMAND_FLAG_GHOST | GAME_COMMAND_FLAG_ALLOW_DURING_PAUSED);
+            auto res = GameActions::Execute(&smallSceneryPlaceAction);
+            if (res.Error != GameActions::Status::Ok)
+                return kMoney64Undefined;
+
+            const auto placementData = res.GetData<SmallSceneryPlaceActionResult>();
+
+            gSceneryPlaceRotation = loc.direction;
+            gSceneryPlaceObject.SceneryType = SCENERY_TYPE_SMALL;
+            gSceneryPlaceObject.EntryIndex = entryIndex;
+
+            gSceneryGhostPosition = { loc, placementData.BaseHeight };
+            gSceneryQuadrant = placementData.SceneryQuadrant;
+            if (placementData.GroundFlags & ELEMENT_IS_UNDERGROUND)
+            {
+                // Set underground on
+                ViewportSetVisibility(ViewportVisibility::UndergroundViewGhostOn);
+            }
+            else
+            {
+                // Set underground off
+                ViewportSetVisibility(ViewportVisibility::UndergroundViewGhostOff);
+            }
+
+            gSceneryGhostType |= SCENERY_GHOST_FLAG_0;
+            return res.Cost;
+        }
+
+        money64 TryPlaceGhostPathAddition(CoordsXYZ loc, ObjectEntryIndex entryIndex)
+        {
+            SceneryRemoveGhostToolPlacement();
+
+            // 6e265b
+            auto footpathAdditionPlaceAction = FootpathAdditionPlaceAction(loc, entryIndex);
+            footpathAdditionPlaceAction.SetFlags(GAME_COMMAND_FLAG_GHOST | GAME_COMMAND_FLAG_ALLOW_DURING_PAUSED);
+            footpathAdditionPlaceAction.SetCallback([=](const GameAction* ga, const GameActions::Result* result) {
+                if (result->Error != GameActions::Status::Ok)
+                {
+                    return;
+                }
+                gSceneryGhostPosition = loc;
+                gSceneryGhostType |= SCENERY_GHOST_FLAG_1;
+            });
+            auto res = GameActions::Execute(&footpathAdditionPlaceAction);
+            if (res.Error != GameActions::Status::Ok)
+                return kMoney64Undefined;
+
+            return res.Cost;
+        }
+
+        money64 TryPlaceGhostWall(
+            CoordsXYZ loc, uint8_t edge, ObjectEntryIndex entryIndex, colour_t primaryColour, colour_t secondaryColour,
+            colour_t tertiaryColour)
+        {
+            SceneryRemoveGhostToolPlacement();
+
+            // 6e26b0
+            auto wallPlaceAction = WallPlaceAction(entryIndex, loc, edge, primaryColour, secondaryColour, tertiaryColour);
+            wallPlaceAction.SetFlags(
+                GAME_COMMAND_FLAG_GHOST | GAME_COMMAND_FLAG_ALLOW_DURING_PAUSED | GAME_COMMAND_FLAG_NO_SPEND);
+            wallPlaceAction.SetCallback([=](const GameAction* ga, const GameActions::Result* result) {
+                if (result->Error != GameActions::Status::Ok)
+                    return;
+
+                const auto placementData = result->GetData<WallPlaceActionResult>();
+                gSceneryGhostPosition = { loc, placementData.BaseHeight };
+                gSceneryGhostWallRotation = edge;
+
+                gSceneryGhostType |= SCENERY_GHOST_FLAG_2;
+            });
+
+            auto res = GameActions::Execute(&wallPlaceAction);
+            if (res.Error != GameActions::Status::Ok)
+                return kMoney64Undefined;
+
+            return res.Cost;
+        }
+
+        money64 TryPlaceGhostLargeScenery(
+            CoordsXYZD loc, ObjectEntryIndex entryIndex, colour_t primaryColour, colour_t secondaryColour,
+            colour_t tertiaryColour)
+        {
+            SceneryRemoveGhostToolPlacement();
+
+            // 6e25a7
+            auto sceneryPlaceAction = LargeSceneryPlaceAction(loc, entryIndex, primaryColour, secondaryColour, tertiaryColour);
+            sceneryPlaceAction.SetFlags(
+                GAME_COMMAND_FLAG_GHOST | GAME_COMMAND_FLAG_ALLOW_DURING_PAUSED | GAME_COMMAND_FLAG_NO_SPEND);
+            auto res = GameActions::Execute(&sceneryPlaceAction);
+            if (res.Error != GameActions::Status::Ok)
+                return kMoney64Undefined;
+
+            const auto placementData = res.GetData<LargeSceneryPlaceActionResult>();
+
+            gSceneryPlaceRotation = loc.direction;
+
+            gSceneryGhostPosition = { loc, placementData.firstTileHeight };
+            if (placementData.GroundFlags & ELEMENT_IS_UNDERGROUND)
+            {
+                // Set underground on
+                ViewportSetVisibility(ViewportVisibility::UndergroundViewGhostOn);
+            }
+            else
+            {
+                // Set underground off
+                ViewportSetVisibility(ViewportVisibility::UndergroundViewGhostOff);
+            }
+
+            gSceneryGhostType |= SCENERY_GHOST_FLAG_3;
+            return res.Cost;
+        }
+
+        money64 TryPlaceGhostBanner(CoordsXYZD loc, ObjectEntryIndex entryIndex)
+        {
+            SceneryRemoveGhostToolPlacement();
+
+            // 6e2612
+            auto primaryColour = gWindowSceneryPrimaryColour;
+            auto bannerPlaceAction = BannerPlaceAction(loc, entryIndex, primaryColour);
+            bannerPlaceAction.SetFlags(
+                GAME_COMMAND_FLAG_GHOST | GAME_COMMAND_FLAG_ALLOW_DURING_PAUSED | GAME_COMMAND_FLAG_NO_SPEND);
+            auto res = GameActions::Execute(&bannerPlaceAction);
+            if (res.Error != GameActions::Status::Ok)
+                return kMoney64Undefined;
+
+            gSceneryGhostPosition = loc;
+            gSceneryGhostPosition.z += PATH_HEIGHT_STEP;
+            gSceneryPlaceRotation = loc.direction;
+            gSceneryGhostType |= SCENERY_GHOST_FLAG_4;
+            return res.Cost;
+        }
+
+        /**
+         *
+         *  rct2: 0x006E3158
+         */
+        void RepaintSceneryToolDown(const ScreenCoordsXY& screenCoords, WidgetIndex widgetIndex)
+        {
+            auto flag = EnumsToFlags(
+                ViewportInteractionItem::Scenery, ViewportInteractionItem::Wall, ViewportInteractionItem::LargeScenery,
+                ViewportInteractionItem::Banner);
+            auto info = GetMapCoordinatesFromPos(screenCoords, flag);
+            switch (info.SpriteType)
+            {
+                case ViewportInteractionItem::Scenery:
+                {
+                    auto* sceneryEntry = info.Element->AsSmallScenery()->GetEntry();
+
+                    // If can't repaint
+                    if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_HAS_PRIMARY_COLOUR | SMALL_SCENERY_FLAG_HAS_GLASS))
+                        return;
+
+                    uint8_t quadrant = info.Element->AsSmallScenery()->GetSceneryQuadrant();
+                    auto repaintScenery = SmallScenerySetColourAction(
+                        { info.Loc, info.Element->GetBaseZ() }, quadrant, info.Element->AsSmallScenery()->GetEntryIndex(),
+                        gWindowSceneryPrimaryColour, gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                    GameActions::Execute(&repaintScenery);
+                    break;
+                }
+                case ViewportInteractionItem::Wall:
+                {
+                    auto* scenery_entry = info.Element->AsWall()->GetEntry();
+
+                    // If can't repaint
+                    if (!(scenery_entry->flags & (WALL_SCENERY_HAS_PRIMARY_COLOUR | WALL_SCENERY_HAS_GLASS)))
+                        return;
+
+                    auto repaintScenery = WallSetColourAction(
+                        { info.Loc, info.Element->GetBaseZ(), info.Element->GetDirection() }, gWindowSceneryPrimaryColour,
+                        gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                    GameActions::Execute(&repaintScenery);
+                    break;
+                }
+                case ViewportInteractionItem::LargeScenery:
+                {
+                    auto* sceneryEntry = info.Element->AsLargeScenery()->GetEntry();
+
+                    // If can't repaint
+                    if (!(sceneryEntry->flags & LARGE_SCENERY_FLAG_HAS_PRIMARY_COLOUR))
+                        return;
+
+                    auto repaintScenery = LargeScenerySetColourAction(
+                        { info.Loc, info.Element->GetBaseZ(), info.Element->GetDirection() },
+                        info.Element->AsLargeScenery()->GetSequenceIndex(), gWindowSceneryPrimaryColour,
+                        gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                    GameActions::Execute(&repaintScenery);
+                    break;
+                }
+                case ViewportInteractionItem::Banner:
+                {
+                    auto banner = info.Element->AsBanner()->GetBanner();
+                    if (banner != nullptr)
+                    {
+                        auto* bannerEntry = OpenRCT2::ObjectManager::GetObjectEntry<BannerSceneryEntry>(banner->type);
+                        if (bannerEntry->flags & BANNER_ENTRY_FLAG_HAS_PRIMARY_COLOUR)
+                        {
+                            auto repaintScenery = BannerSetColourAction(
+                                { info.Loc, info.Element->GetBaseZ(), info.Element->AsBanner()->GetPosition() },
+                                gWindowSceneryPrimaryColour);
+
+                            GameActions::Execute(&repaintScenery);
+                        }
+                    }
+                    break;
+                }
+                default:
+                    return;
+            }
+        }
+
+        void SceneryEyedropperToolDown(const ScreenCoordsXY& screenCoords, WidgetIndex widgetIndex)
+        {
+            auto flag = EnumsToFlags(
+                ViewportInteractionItem::Scenery, ViewportInteractionItem::Wall, ViewportInteractionItem::LargeScenery,
+                ViewportInteractionItem::Banner, ViewportInteractionItem::PathAddition);
+            auto info = GetMapCoordinatesFromPos(screenCoords, flag);
+            switch (info.SpriteType)
+            {
+                case ViewportInteractionItem::Scenery:
+                {
+                    SmallSceneryElement* sceneryElement = info.Element->AsSmallScenery();
+                    auto entryIndex = sceneryElement->GetEntryIndex();
+                    auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<SmallSceneryEntry>(entryIndex);
+                    if (sceneryEntry != nullptr)
+                    {
+                        WindowScenerySetSelectedItem(
+                            { SCENERY_TYPE_SMALL, entryIndex }, sceneryElement->GetPrimaryColour(),
+                            sceneryElement->GetSecondaryColour(), sceneryElement->GetTertiaryColour(),
+                            sceneryElement->GetDirectionWithOffset(GetCurrentRotation()));
+                    }
+                    break;
+                }
+                case ViewportInteractionItem::Wall:
+                {
+                    auto entryIndex = info.Element->AsWall()->GetEntryIndex();
+                    auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<WallSceneryEntry>(entryIndex);
+                    if (sceneryEntry != nullptr)
+                    {
+                        WindowScenerySetSelectedItem(
+                            { SCENERY_TYPE_WALL, entryIndex }, info.Element->AsWall()->GetPrimaryColour(),
+                            info.Element->AsWall()->GetSecondaryColour(), info.Element->AsWall()->GetTertiaryColour(),
+                            std::nullopt);
+                    }
+                    break;
+                }
+                case ViewportInteractionItem::LargeScenery:
+                {
+                    auto entryIndex = info.Element->AsLargeScenery()->GetEntryIndex();
+                    auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<LargeSceneryEntry>(entryIndex);
+                    if (sceneryEntry != nullptr)
+                    {
+                        WindowScenerySetSelectedItem(
+                            { SCENERY_TYPE_LARGE, entryIndex }, info.Element->AsLargeScenery()->GetPrimaryColour(),
+                            info.Element->AsLargeScenery()->GetSecondaryColour(), std::nullopt,
+                            (GetCurrentRotation() + info.Element->GetDirection()) & 3);
+                    }
+                    break;
+                }
+                case ViewportInteractionItem::Banner:
+                {
+                    auto banner = info.Element->AsBanner()->GetBanner();
+                    if (banner != nullptr)
+                    {
+                        auto sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<BannerSceneryEntry>(banner->type);
+                        if (sceneryEntry != nullptr)
+                        {
+                            WindowScenerySetSelectedItem(
+                                { SCENERY_TYPE_BANNER, banner->type }, std::nullopt, std::nullopt, std::nullopt, std::nullopt);
+                        }
+                    }
+                    break;
+                }
+                case ViewportInteractionItem::PathAddition:
+                {
+                    auto entryIndex = info.Element->AsPath()->GetAdditionEntryIndex();
+                    auto* pathAdditionEntry = OpenRCT2::ObjectManager::GetObjectEntry<PathAdditionEntry>(entryIndex);
+                    if (pathAdditionEntry != nullptr)
+                    {
+                        WindowScenerySetSelectedItem(
+                            { SCENERY_TYPE_PATH_ITEM, entryIndex }, std::nullopt, std::nullopt, std::nullopt, std::nullopt);
+                    }
+                    break;
+                }
+                default:
+                    break;
+            }
+        }
+
+        void Sub6E1F34UpdateScreenCoordsAndButtonsPressed(bool canRaiseItem, ScreenCoordsXY& screenPos)
+        {
+            if (!canRaiseItem && !GetGameState().Cheats.DisableSupportLimits)
+            {
+                gSceneryCtrlPressed = false;
+                gSceneryShiftPressed = false;
+            }
+            else
+            {
+                if (!gSceneryCtrlPressed)
+                {
+                    if (InputTestPlaceObjectModifier(PLACE_OBJECT_MODIFIER_COPY_Z))
+                    {
+                        // CTRL pressed
+                        constexpr auto flag = EnumsToFlags(
+                            ViewportInteractionItem::Terrain, ViewportInteractionItem::Ride, ViewportInteractionItem::Scenery,
+                            ViewportInteractionItem::Footpath, ViewportInteractionItem::Wall,
+                            ViewportInteractionItem::LargeScenery);
+                        auto info = GetMapCoordinatesFromPos(screenPos, flag);
+
+                        if (info.SpriteType != ViewportInteractionItem::None)
+                        {
+                            gSceneryCtrlPressed = true;
+                            gSceneryCtrlPressZ = info.Element->GetBaseZ();
+                        }
+                    }
+                }
+                else
+                {
+                    if (!(InputTestPlaceObjectModifier(PLACE_OBJECT_MODIFIER_COPY_Z)))
+                    {
+                        // CTRL not pressed
+                        gSceneryCtrlPressed = false;
+                    }
+                }
+
+                if (!gSceneryShiftPressed)
+                {
+                    if (InputTestPlaceObjectModifier(PLACE_OBJECT_MODIFIER_SHIFT_Z))
+                    {
+                        // SHIFT pressed
+                        gSceneryShiftPressed = true;
+                        gSceneryShiftPressX = screenPos.x;
+                        gSceneryShiftPressY = screenPos.y;
+                        gSceneryShiftPressZOffset = 0;
+                    }
+                }
+                else
+                {
+                    if (InputTestPlaceObjectModifier(PLACE_OBJECT_MODIFIER_SHIFT_Z))
+                    {
+                        // SHIFT pressed
+                        gSceneryShiftPressZOffset = (gSceneryShiftPressY - screenPos.y + 4);
+                        // Scale delta by zoom to match mouse position.
+                        auto* mainWnd = WindowGetMain();
+                        if (mainWnd != nullptr && mainWnd->viewport != nullptr)
+                        {
+                            gSceneryShiftPressZOffset = mainWnd->viewport->zoom.ApplyTo(gSceneryShiftPressZOffset);
+                        }
+                        gSceneryShiftPressZOffset = Floor2(gSceneryShiftPressZOffset, 8);
+
+                        screenPos.x = gSceneryShiftPressX;
+                        screenPos.y = gSceneryShiftPressY;
+                    }
+                    else
+                    {
+                        // SHIFT not pressed
+                        gSceneryShiftPressed = false;
+                    }
+                }
+            }
+        }
+
+        void Sub6E1F34SmallScenery(
+            const ScreenCoordsXY& sourceScreenPos, ObjectEntryIndex sceneryIndex, CoordsXY& gridPos, uint8_t* outQuadrant,
+            Direction* outRotation)
+        {
+            auto* w = WindowFindByClass(WindowClass::Scenery);
+
+            if (w == nullptr)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            auto screenPos = sourceScreenPos;
+            uint16_t maxPossibleHeight = ZoomLevel::max().ApplyTo(
+                std::numeric_limits<decltype(TileElement::BaseHeight)>::max() - 32);
+            bool can_raise_item = false;
+
+            const auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<SmallSceneryEntry>(sceneryIndex);
+            if (sceneryEntry == nullptr)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            maxPossibleHeight -= sceneryEntry->height;
+            if (sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_STACKABLE))
+            {
+                can_raise_item = true;
+            }
+
+            Sub6E1F34UpdateScreenCoordsAndButtonsPressed(can_raise_item, screenPos);
+
+            // Small scenery
+            if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_FULL_TILE))
+            {
+                uint8_t quadrant = 0;
+
+                // If CTRL not pressed
+                if (!gSceneryCtrlPressed)
+                {
+                    auto gridCoords = ScreenGetMapXYQuadrant(screenPos, &quadrant);
+                    if (!gridCoords.has_value())
+                    {
+                        gridPos.SetNull();
+                        return;
+                    }
+                    gridPos = gridCoords.value();
+
+                    gSceneryPlaceZ = 0;
+
+                    // If SHIFT pressed
+                    if (gSceneryShiftPressed)
+                    {
+                        auto* surfaceElement = MapGetSurfaceElementAt(gridPos);
+
+                        if (surfaceElement == nullptr)
+                        {
+                            gridPos.SetNull();
+                            return;
+                        }
+
+                        int16_t z = (surfaceElement->GetBaseZ()) & 0xFFF0;
+                        z += gSceneryShiftPressZOffset;
+
+                        z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                        gSceneryPlaceZ = z;
+                    }
+                }
+                else
+                {
+                    int16_t z = gSceneryCtrlPressZ;
+
+                    auto mapCoords = ScreenGetMapXYQuadrantWithZ(screenPos, z, &quadrant);
+                    if (!mapCoords.has_value())
+                    {
+                        gridPos.SetNull();
+                        return;
+                    }
+                    gridPos = mapCoords.value();
+
+                    // If SHIFT pressed
+                    if (gSceneryShiftPressed)
+                    {
+                        z += gSceneryShiftPressZOffset;
+                    }
+
+                    z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                    gSceneryPlaceZ = z;
+                }
+
+                if (gridPos.IsNull())
+                    return;
+
+                uint8_t rotation = gWindowSceneryRotation;
+
+                if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_ROTATABLE))
+                {
+                    rotation = UtilRand() & 0xFF;
+                }
+
+                rotation -= GetCurrentRotation();
+                rotation &= 0x3;
+
+                if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+                {
+                    VirtualFloorSetHeight(gSceneryPlaceZ);
+                }
+
+                *outQuadrant = quadrant ^ 2;
+                *outRotation = rotation;
+
+                return;
+            }
+
+            // If CTRL not pressed
+            if (!gSceneryCtrlPressed)
+            {
+                constexpr auto flag = EnumsToFlags(ViewportInteractionItem::Terrain, ViewportInteractionItem::Water);
+
+                auto info = GetMapCoordinatesFromPos(screenPos, flag);
+                gridPos = info.Loc;
+
+                if (info.SpriteType == ViewportInteractionItem::None)
+                {
+                    gridPos.SetNull();
+                    return;
+                }
+
+                // If CTRL and SHIFT not pressed
+                gSceneryPlaceZ = 0;
+
+                // If SHIFT pressed
+                if (gSceneryShiftPressed)
+                {
+                    auto surfaceElement = MapGetSurfaceElementAt(gridPos);
+
+                    if (surfaceElement == nullptr)
+                    {
+                        gridPos.SetNull();
+                        return;
+                    }
+
+                    int16_t z = (surfaceElement->GetBaseZ()) & 0xFFF0;
+                    z += gSceneryShiftPressZOffset;
+
+                    z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                    gSceneryPlaceZ = z;
+                }
+            }
+            else
+            {
+                int16_t z = gSceneryCtrlPressZ;
+                auto coords = ScreenGetMapXYWithZ(screenPos, z);
+                if (coords.has_value())
+                {
+                    gridPos = *coords;
+                }
+                else
+                {
+                    gridPos.SetNull();
+                }
+                // If SHIFT pressed
+                if (gSceneryShiftPressed)
+                {
+                    z += gSceneryShiftPressZOffset;
+                }
+
+                z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                gSceneryPlaceZ = z;
+            }
+
+            if (gridPos.IsNull())
+                return;
+
+            gridPos = gridPos.ToTileStart();
+            uint8_t rotation = gWindowSceneryRotation;
+
+            if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_ROTATABLE))
+            {
+                rotation = UtilRand() & 0xFF;
+            }
+
+            rotation -= GetCurrentRotation();
+            rotation &= 0x3;
+
+            if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+            {
+                VirtualFloorSetHeight(gSceneryPlaceZ);
+            }
+
+            *outQuadrant = 0;
+            *outRotation = rotation;
+        }
+
+        void Sub6E1F34PathItem(
+            const ScreenCoordsXY& sourceScreenPos, ObjectEntryIndex sceneryIndex, CoordsXY& gridPos, int32_t* outZ)
+        {
+            auto* w = WindowFindByClass(WindowClass::Scenery);
+
+            if (w == nullptr)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            auto screenPos = sourceScreenPos;
+            Sub6E1F34UpdateScreenCoordsAndButtonsPressed(false, screenPos);
+
+            // Path bits
+            constexpr auto flag = EnumsToFlags(ViewportInteractionItem::Footpath, ViewportInteractionItem::PathAddition);
+            auto info = GetMapCoordinatesFromPos(screenPos, flag);
+            gridPos = info.Loc;
+
+            if (info.SpriteType == ViewportInteractionItem::None)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+            {
+                VirtualFloorSetHeight(gSceneryPlaceZ);
+            }
+
+            *outZ = info.Element->GetBaseZ();
+        }
+
+        void Sub6E1F34Wall(
+            const ScreenCoordsXY& sourceScreenPos, ObjectEntryIndex sceneryIndex, CoordsXY& gridPos, uint8_t* outEdges)
+        {
+            auto* w = WindowFindByClass(WindowClass::Scenery);
+
+            if (w == nullptr)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            auto screenPos = sourceScreenPos;
+            uint16_t maxPossibleHeight = ZoomLevel::max().ApplyTo(
+                std::numeric_limits<decltype(TileElement::BaseHeight)>::max() - 32);
+
+            auto* wallEntry = OpenRCT2::ObjectManager::GetObjectEntry<WallSceneryEntry>(sceneryIndex);
+            if (wallEntry != nullptr)
+            {
+                maxPossibleHeight -= wallEntry->height;
+            }
+
+            Sub6E1F34UpdateScreenCoordsAndButtonsPressed(true, screenPos);
+
+            // Walls
+            uint8_t edge;
+            // If CTRL not pressed
+            if (!gSceneryCtrlPressed)
+            {
+                auto gridCoords = ScreenGetMapXYSide(screenPos, &edge);
+                if (!gridCoords.has_value())
+                {
+                    gridPos.SetNull();
+                    return;
+                }
+                gridPos = gridCoords.value();
+
+                gSceneryPlaceZ = 0;
+
+                // If SHIFT pressed
+                if (gSceneryShiftPressed)
+                {
+                    auto* surfaceElement = MapGetSurfaceElementAt(gridPos);
+
+                    if (surfaceElement == nullptr)
+                    {
+                        gridPos.SetNull();
+                        return;
+                    }
+
+                    int16_t z = (surfaceElement->GetBaseZ()) & 0xFFF0;
+                    z += gSceneryShiftPressZOffset;
+
+                    z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                    gSceneryPlaceZ = z;
+                }
+            }
+            else
+            {
+                int16_t z = gSceneryCtrlPressZ;
+                auto mapCoords = ScreenGetMapXYSideWithZ(screenPos, z, &edge);
+                if (!mapCoords.has_value())
+                {
+                    gridPos.SetNull();
+                    return;
+                }
+                gridPos = mapCoords.value();
+
+                // If SHIFT pressed
+                if (gSceneryShiftPressed)
+                {
+                    z += gSceneryShiftPressZOffset;
+                }
+
+                z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                gSceneryPlaceZ = z;
+            }
+
+            if (gridPos.IsNull())
+                return;
+
+            if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+            {
+                VirtualFloorSetHeight(gSceneryPlaceZ);
+            }
+
+            *outEdges = edge;
+        }
+
+        void Sub6E1F34LargeScenery(
+            const ScreenCoordsXY& sourceScreenPos, ObjectEntryIndex sceneryIndex, CoordsXY& gridPos, Direction* outDirection)
+        {
+            auto* w = WindowFindByClass(WindowClass::Scenery);
+
+            if (w == nullptr)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            auto screenPos = sourceScreenPos;
+            uint16_t maxPossibleHeight = ZoomLevel::max().ApplyTo(
+                std::numeric_limits<decltype(TileElement::BaseHeight)>::max() - 32);
+
+            auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<LargeSceneryEntry>(sceneryIndex);
+            if (sceneryEntry)
+            {
+                int16_t maxClearZ = 0;
+                for (int32_t i = 0; sceneryEntry->tiles[i].x_offset != -1; ++i)
+                {
+                    maxClearZ = std::max<int16_t>(maxClearZ, sceneryEntry->tiles[i].z_clearance);
+                }
+                maxPossibleHeight = std::max(0, maxPossibleHeight - maxClearZ);
+            }
+
+            Sub6E1F34UpdateScreenCoordsAndButtonsPressed(true, screenPos);
+
+            // Large scenery
+            // If CTRL not pressed
+            if (!gSceneryCtrlPressed)
+            {
+                const CoordsXY mapCoords = ViewportInteractionGetTileStartAtCursor(screenPos);
+                gridPos = mapCoords;
+
+                if (gridPos.IsNull())
+                    return;
+
+                gSceneryPlaceZ = 0;
+
+                // If SHIFT pressed
+                if (gSceneryShiftPressed)
+                {
+                    auto* surfaceElement = MapGetSurfaceElementAt(gridPos);
+
+                    if (surfaceElement == nullptr)
+                    {
+                        gridPos.SetNull();
+                        return;
+                    }
+
+                    int16_t z = (surfaceElement->GetBaseZ()) & 0xFFF0;
+                    z += gSceneryShiftPressZOffset;
+
+                    z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                    gSceneryPlaceZ = z;
+                }
+            }
+            else
+            {
+                int16_t z = gSceneryCtrlPressZ;
+                auto coords = ScreenGetMapXYWithZ(screenPos, z);
+                if (coords.has_value())
+                {
+                    gridPos = *coords;
+                }
+                else
+                {
+                    gridPos.SetNull();
+                }
+
+                // If SHIFT pressed
+                if (gSceneryShiftPressed)
+                {
+                    z += gSceneryShiftPressZOffset;
+                }
+
+                z = std::clamp<int16_t>(z, 16, maxPossibleHeight);
+
+                gSceneryPlaceZ = z;
+            }
+
+            if (gridPos.IsNull())
+                return;
+
+            gridPos = gridPos.ToTileStart();
+
+            Direction rotation = gWindowSceneryRotation;
+            rotation -= GetCurrentRotation();
+            rotation &= 0x3;
+
+            if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+            {
+                VirtualFloorSetHeight(gSceneryPlaceZ);
+            }
+
+            *outDirection = rotation;
+        }
+
+        void Sub6E1F34Banner(
+            const ScreenCoordsXY& sourceScreenPos, ObjectEntryIndex sceneryIndex, CoordsXY& gridPos, int32_t* outZ,
+            Direction* outDirection)
+        {
+            auto* w = WindowFindByClass(WindowClass::Scenery);
+
+            if (w == nullptr)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            auto screenPos = sourceScreenPos;
+            Sub6E1F34UpdateScreenCoordsAndButtonsPressed(false, screenPos);
+
+            // Banner
+            constexpr auto flag = EnumsToFlags(ViewportInteractionItem::Footpath, ViewportInteractionItem::PathAddition);
+            auto info = GetMapCoordinatesFromPos(screenPos, flag);
+            gridPos = info.Loc;
+
+            if (info.SpriteType == ViewportInteractionItem::None)
+            {
+                gridPos.SetNull();
+                return;
+            }
+
+            uint8_t rotation = gWindowSceneryRotation;
+            rotation -= GetCurrentRotation();
+            rotation &= 0x3;
+
+            auto z = info.Element->GetBaseZ();
+
+            if (info.Element->AsPath()->IsSloped())
+            {
+                if (rotation != DirectionReverse(info.Element->AsPath()->GetSlopeDirection()))
+                {
+                    z += (2 * kCoordsZStep);
+                }
+            }
+
+            if (Config::Get().general.VirtualFloorStyle != VirtualFloorStyles::Off)
+            {
+                VirtualFloorSetHeight(gSceneryPlaceZ);
+            }
+
+            *outDirection = rotation;
+            *outZ = z;
+        }
+
+        /**
+         *
+         *  rct2: 0x006E2CC6
+         */
+        void SceneryToolDown(const ScreenCoordsXY& screenCoords, WidgetIndex widgetIndex)
+        {
+            SceneryRemoveGhostToolPlacement();
+            if (gWindowSceneryPaintEnabled & 1)
+            {
+                RepaintSceneryToolDown(screenCoords, widgetIndex);
+                return;
+            }
+
+            if (gWindowSceneryEyedropperEnabled)
+            {
+                SceneryEyedropperToolDown(screenCoords, widgetIndex);
+                return;
+            }
+
+            auto selectedTab = WindowSceneryGetTabSelection();
+            uint8_t sceneryType = selectedTab.SceneryType;
+            uint16_t selectedScenery = selectedTab.EntryIndex;
+            CoordsXY gridPos;
+
+            switch (sceneryType)
+            {
+                case SCENERY_TYPE_SMALL:
+                {
+                    uint8_t quadrant;
+                    Direction rotation;
+                    Sub6E1F34SmallScenery(screenCoords, selectedScenery, gridPos, &quadrant, &rotation);
+                    if (gridPos.IsNull())
+                        return;
+
+                    int32_t quantity = 1;
+                    bool isCluster = gWindowSceneryScatterEnabled
+                        && (NetworkGetMode() != NETWORK_MODE_CLIENT
+                            || NetworkCanPerformCommand(NetworkGetCurrentPlayerGroupIndex(), -2));
+
+                    if (isCluster)
+                    {
+                        switch (gWindowSceneryScatterDensity)
+                        {
+                            case ScatterToolDensity::LowDensity:
+                                quantity = gWindowSceneryScatterSize;
+                                break;
+
+                            case ScatterToolDensity::MediumDensity:
+                                quantity = gWindowSceneryScatterSize * 2;
+                                break;
+
+                            case ScatterToolDensity::HighDensity:
+                                quantity = gWindowSceneryScatterSize * 3;
+                                break;
+                        }
+                    }
+
+                    bool forceError = true;
+                    for (int32_t q = 0; q < quantity; q++)
+                    {
+                        int32_t zCoordinate = gSceneryPlaceZ;
+                        auto* sceneryEntry = OpenRCT2::ObjectManager::GetObjectEntry<SmallSceneryEntry>(selectedScenery);
+
+                        int16_t cur_grid_x = gridPos.x;
+                        int16_t cur_grid_y = gridPos.y;
+
+                        if (isCluster)
+                        {
+                            if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_FULL_TILE))
+                            {
+                                quadrant = UtilRand() & 3;
+                            }
+
+                            int16_t grid_x_offset = (UtilRand() % gWindowSceneryScatterSize) - (gWindowSceneryScatterSize / 2);
+                            int16_t grid_y_offset = (UtilRand() % gWindowSceneryScatterSize) - (gWindowSceneryScatterSize / 2);
+                            if (gWindowSceneryScatterSize % 2 == 0)
+                            {
+                                grid_x_offset += 1;
+                                grid_y_offset += 1;
+                            }
+                            cur_grid_x += grid_x_offset * kCoordsXYStep;
+                            cur_grid_y += grid_y_offset * kCoordsXYStep;
+
+                            if (!sceneryEntry->HasFlag(SMALL_SCENERY_FLAG_ROTATABLE))
+                            {
+                                gSceneryPlaceRotation = (gSceneryPlaceRotation + 1) & 3;
+                            }
+                        }
+
+                        uint8_t zAttemptRange = 1;
+                        if (gSceneryPlaceZ != 0 && gSceneryShiftPressed)
+                        {
+                            zAttemptRange = 20;
+                        }
+
+                        auto success = GameActions::Status::Unknown;
+                        // Try find a valid z coordinate
+                        for (; zAttemptRange != 0; zAttemptRange--)
+                        {
+                            auto smallSceneryPlaceAction = SmallSceneryPlaceAction(
+                                { cur_grid_x, cur_grid_y, gSceneryPlaceZ, gSceneryPlaceRotation }, quadrant, selectedScenery,
+                                gWindowSceneryPrimaryColour, gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+                            auto res = GameActions::Query(&smallSceneryPlaceAction);
+                            success = res.Error;
+                            if (res.Error == GameActions::Status::Ok)
+                            {
+                                break;
+                            }
+
+                            if (res.Error == GameActions::Status::InsufficientFunds)
+                            {
+                                break;
+                            }
+                            if (zAttemptRange != 1)
+                            {
+                                gSceneryPlaceZ += 8;
+                            }
+                        }
+
+                        // Actually place
+                        if (success == GameActions::Status::Ok || ((q + 1 == quantity) && forceError))
+                        {
+                            auto smallSceneryPlaceAction = SmallSceneryPlaceAction(
+                                { cur_grid_x, cur_grid_y, gSceneryPlaceZ, gSceneryPlaceRotation }, quadrant, selectedScenery,
+                                gWindowSceneryPrimaryColour, gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                            smallSceneryPlaceAction.SetCallback([=](const GameAction* ga, const GameActions::Result* result) {
+                                if (result->Error == GameActions::Status::Ok)
+                                {
+                                    OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::PlaceItem, result->Position);
+                                }
+                            });
+                            auto res = GameActions::Execute(&smallSceneryPlaceAction);
+                            if (res.Error == GameActions::Status::Ok)
+                            {
+                                forceError = false;
+                            }
+
+                            if (res.Error == GameActions::Status::InsufficientFunds)
+                            {
+                                break;
+                            }
+                        }
+                        gSceneryPlaceZ = zCoordinate;
+                    }
+                    break;
+                }
+                case SCENERY_TYPE_PATH_ITEM:
+                {
+                    int32_t z;
+                    Sub6E1F34PathItem(screenCoords, selectedScenery, gridPos, &z);
+                    if (gridPos.IsNull())
+                        return;
+
+                    auto footpathAdditionPlaceAction = FootpathAdditionPlaceAction({ gridPos, z }, selectedScenery);
+
+                    footpathAdditionPlaceAction.SetCallback([](const GameAction* ga, const GameActions::Result* result) {
+                        if (result->Error != GameActions::Status::Ok)
+                        {
+                            return;
+                        }
+                        OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::PlaceItem, result->Position);
+                    });
+                    auto res = GameActions::Execute(&footpathAdditionPlaceAction);
+                    break;
+                }
+                case SCENERY_TYPE_WALL:
+                {
+                    uint8_t edges;
+                    Sub6E1F34Wall(screenCoords, selectedScenery, gridPos, &edges);
+                    if (gridPos.IsNull())
+                        return;
+
+                    uint8_t zAttemptRange = 1;
+                    if (gSceneryPlaceZ != 0 && gSceneryShiftPressed)
+                    {
+                        zAttemptRange = 20;
+                    }
+
+                    for (; zAttemptRange != 0; zAttemptRange--)
+                    {
+                        auto wallPlaceAction = WallPlaceAction(
+                            selectedScenery, { gridPos, gSceneryPlaceZ }, edges, gWindowSceneryPrimaryColour,
+                            gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                        auto res = GameActions::Query(&wallPlaceAction);
+                        if (res.Error == GameActions::Status::Ok)
+                        {
+                            break;
+                        }
+
+                        if (const auto* message = std::get_if<StringId>(&res.ErrorMessage))
+                        {
+                            if (*message == STR_NOT_ENOUGH_CASH_REQUIRES || *message == STR_CAN_ONLY_BUILD_THIS_ON_WATER)
+                            {
+                                break;
+                            }
+                        }
+
+                        if (zAttemptRange != 1)
+                        {
+                            gSceneryPlaceZ += 8;
+                        }
+                    }
+
+                    auto wallPlaceAction = WallPlaceAction(
+                        selectedScenery, { gridPos, gSceneryPlaceZ }, edges, gWindowSceneryPrimaryColour,
+                        gWindowScenerySecondaryColour, gWindowSceneryTertiaryColour);
+
+                    wallPlaceAction.SetCallback([](const GameAction* ga, const GameActions::Result* result) {
+                        if (result->Error == GameActions::Status::Ok)
+                        {
+                            OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::PlaceItem, result->Position);
+                        }
+                    });
+                    auto res = GameActions::Execute(&wallPlaceAction);
+                    break;
+                }
+                case SCENERY_TYPE_LARGE:
+                {
+                    Direction direction;
+                    Sub6E1F34LargeScenery(screenCoords, selectedScenery, gridPos, &direction);
+                    if (gridPos.IsNull())
+                        return;
+
+                    uint8_t zAttemptRange = 1;
+                    if (gSceneryPlaceZ != 0 && gSceneryShiftPressed)
+                    {
+                        zAttemptRange = 20;
+                    }
+
+                    for (; zAttemptRange != 0; zAttemptRange--)
+                    {
+                        CoordsXYZD loc = { gridPos, gSceneryPlaceZ, direction };
+
+                        auto sceneryPlaceAction = LargeSceneryPlaceAction(
+                            loc, selectedScenery, gWindowSceneryPrimaryColour, gWindowScenerySecondaryColour,
+                            gWindowSceneryTertiaryColour);
+
+                        auto res = GameActions::Query(&sceneryPlaceAction);
+                        if (res.Error == GameActions::Status::Ok)
+                        {
+                            break;
+                        }
+
+                        if (const auto* message = std::get_if<StringId>(&res.ErrorMessage))
+                        {
+                            if (*message == STR_NOT_ENOUGH_CASH_REQUIRES || *message == STR_CAN_ONLY_BUILD_THIS_ON_WATER)
+                            {
+                                break;
+                            }
+                        }
+
+                        if (zAttemptRange != 1)
+                        {
+                            gSceneryPlaceZ += 8;
+                        }
+                    }
+
+                    CoordsXYZD loc = { gridPos, gSceneryPlaceZ, direction };
+
+                    auto sceneryPlaceAction = LargeSceneryPlaceAction(
+                        loc, selectedScenery, gWindowSceneryPrimaryColour, gWindowScenerySecondaryColour,
+                        gWindowSceneryTertiaryColour);
+                    sceneryPlaceAction.SetCallback([=](const GameAction* ga, const GameActions::Result* result) {
+                        if (result->Error == GameActions::Status::Ok)
+                        {
+                            OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::PlaceItem, result->Position);
+                        }
+                        else
+                        {
+                            OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::Error, { loc.x, loc.y, gSceneryPlaceZ });
+                        }
+                    });
+                    auto res = GameActions::Execute(&sceneryPlaceAction);
+                    break;
+                }
+                case SCENERY_TYPE_BANNER:
+                {
+                    int32_t z;
+                    Direction direction;
+                    Sub6E1F34Banner(screenCoords, selectedScenery, gridPos, &z, &direction);
+                    if (gridPos.IsNull())
+                        return;
+
+                    CoordsXYZD loc{ gridPos, z, direction };
+                    auto primaryColour = gWindowSceneryPrimaryColour;
+                    auto bannerPlaceAction = BannerPlaceAction(loc, selectedScenery, primaryColour);
+                    bannerPlaceAction.SetCallback([=](const GameAction* ga, const GameActions::Result* result) {
+                        if (result->Error == GameActions::Status::Ok)
+                        {
+                            auto data = result->GetData<BannerPlaceActionResult>();
+                            OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::PlaceItem, result->Position);
+                            ContextOpenDetailWindow(WD_BANNER, data.bannerId.ToUnderlying());
+                        }
+                    });
+                    GameActions::Execute(&bannerPlaceAction);
+                    break;
+                }
+            }
+        }
     };
 
     WindowBase* SceneryOpen()
@@ -1753,6 +3260,21 @@ static Widget WindowSceneryBaseWidgets[] = {
         if (w != nullptr)
         {
             w->Init();
+        }
+    }
+
+    void ToggleSceneryWindow()
+    {
+        if ((InputTestFlag(INPUT_FLAG_TOOL_ACTIVE)) && gCurrentToolWidget.window_classification == WindowClass::Scenery
+            && gCurrentToolWidget.widget_index == WIDX_SCENERY_BACKGROUND)
+        {
+            ToolCancel();
+        }
+        else
+        {
+            auto* toolWindow = ContextOpenWindow(WindowClass::Scenery);
+            ToolSet(*toolWindow, WIDX_SCENERY_BACKGROUND, Tool::Arrow);
+            InputSetFlag(INPUT_FLAG_6, true);
         }
     }
 } // namespace OpenRCT2::Ui::Windows
