+++ src/openrct2/interface/Window.h
@@ -355,6 +355,7 @@ constexpr int32_t WC_RIDE_CONSTRUCTION__WIDX_ROTATE = 32;
 constexpr int32_t WC_MAZE_CONSTRUCTION__WIDX_MAZE_DIRECTION_GROUPBOX = WC_RIDE_CONSTRUCTION__WIDX_CONSTRUCT;
 constexpr int32_t WC_MAZE_CONSTRUCTION__WIDX_MAZE_ENTRANCE = WC_RIDE_CONSTRUCTION__WIDX_ENTRANCE;
 constexpr int32_t WC_MAZE_CONSTRUCTION__WIDX_MAZE_EXIT = WC_RIDE_CONSTRUCTION__WIDX_EXIT;
+constexpr int32_t WC_SCENERY__WIDX_SCENERY_BACKGROUND = 0;
 constexpr int32_t WC_SCENERY__WIDX_SCENERY_TAB_1 = 15;
 constexpr int32_t WC_SCENERY__WIDX_SCENERY_ROTATE_OBJECTS_BUTTON = 5;
 constexpr int32_t WC_SCENERY__WIDX_SCENERY_EYEDROPPER_BUTTON = 10;
